Made plugins a directory module
